renames testimonial to twittercard
